отправка на первое ревью пр11

+++ src/utils/Api.js
@@ -56,27 +56,26 @@
       headers: this._headers,
     });
   }
-
-  likeCard(id) {
-    return this._request(`${this._baseUrl}/cards/${id}/likes`, {
-      method: "PUT",
-      headers: this._headers,
-    });
+  changeLikeCardStatus(id,isLiked) {
+    if (isLiked) {
+      return this._request(`${this._baseUrl}/cards/${id}/likes`, {
+        method: "PUT",
+        headers: this._headers,
+      });
+    } else {
+      return this._request(`${this._baseUrl}/cards/${id}/likes`, {
+        method: "DELETE",
+        headers: this._headers,
+      });
+    }
   }
-
-  dislikeCard(id) {
-    return this._request(`${this._baseUrl}/cards/${id}/likes`, {
-      method: "DELETE",
-      headers: this._headers,
-    });
-  }
-
+  
   setAvatar(data) {
     return this._request(`${this._baseUrl}/users/me/avatar`, {
       method: "PATCH",
       headers: this._headers,
       body: JSON.stringify({
-        avatar: data.link,
+        avatar: data.avatar,
       }),
     });
   }
